# frontend/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "dgat-backend",
        "description": "",
        "license": {
            "name": ""
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "https://app.decidel.app/api",
            "description": "DGAT Backend Server"
        }
    ],
    "paths": {
...
        "/dimensions/{id}/recommendations": {
            "get": {
                "tags": [
                    "recommendations"
                ],
                "summary": "List recommendations by dimension",
                "description": "Retrieves a paginated list of recommendations for a specific dimension.",
                "operationId": "list_recommendations_by_dimension",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Dimension ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (starts from 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "nullable": true
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Number of items per page (default: 10, max: 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "nullable": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of recommendations for the dimension",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePaginatedRecommendationResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/recommendations": {
            "get": {
                "tags": [
                    "recommendations"
                ],
                "summary": "List all recommendations with pagination",
                "description": "Retrieves a paginated list of all recommendations in the system.",
                "operationId": "list_recommendations",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number (starts from 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "nullable": true
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Number of items per page (default: 10, max: 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "nullable": true
                        }
                    },
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Language code to filter content (e.g. en, fr, pt, ss). Defaults to en.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "en"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of recommendations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePaginatedRecommendationResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "recommendations"
                ],
                "summary": "Create a new recommendation",
                "description": "Creates a new recommendation with the provided details.",
                "operationId": "create_recommendation",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRecommendationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Recommendation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseRecommendationResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/recommendations/{id}": {
            "get": {
                "tags": [
                    "recommendations"
                ],
                "summary": "Get a specific recommendation by ID",
                "description": "Retrieves the details of a specific recommendation using its unique identifier.",
                "operationId": "get_recommendation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Recommendation ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recommendation found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseRecommendationResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recommendation not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "recommendations"
                ],
                "summary": "Update an existing recommendation",
                "description": "Updates the priority and/or description of an existing recommendation.",
                "operationId": "update_recommendation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Recommendation ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRecommendationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Recommendation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseRecommendationResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "404": {
                        "description": "Recommendation not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "recommendations"
                ],
                "summary": "Delete a recommendation",
                "description": "Deletes a recommendation by its ID.",
                "operationId": "delete_recommendation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Recommendation ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recommendation deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseEmpty"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recommendation not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponsePaginatedRecommendationResponse": {
                "type": "object",
                "description": "Standard API response wrapper",
                "required": [
                    "success"
                ],
                "properties": {
                    "data": {
                        "allOf": [
                            {
                                "type": "object",
                                "description": "Paginated response",
                                "required": [
                                    "items",
                                    "total",
                                    "page",
                                    "limit",
                                    "total_pages"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "required": [
                                                "recommendation_id",
                                                "dimension_id",
                                                "priority",
                                                "description",
                                                "created_at",
                                                "updated_at"
                                            ],
                                            "properties": {
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "description": "When the recommendation was created",
                                                    "example": "2023-01-01T00:00:00Z"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "description": "Detailed description of the recommendation",
                                                    "example": "Implement automated testing for critical components"
                                                },
                                                "dimension_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "description": "Reference to the dimension",
                                                    "example": "550e8400-e29b-41d4-a716-446655440002"
                                                },
                                                "priority": {
                                                    "$ref": "#/components/schemas/RecommendationPriority"
                                                },
                                                "recommendation_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "description": "Unique identifier for the recommendation",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "description": "When the recommendation was last updated",
                                                    "example": "2023-01-01T00:00:00Z"
                                                }
                                            },
                                            "example": {
                                                "recommendation_id": "550e8400-e29b-41d4-a716-446655440000",
                                                "dimension_id": "550e8400-e29b-41d4-a716-446655440002",
                                                "priority": "HIGH",
                                                "description": "Implement automated testing for critical components",
                                                "created_at": "2023-01-01T00:00:00Z",
                                                "updated_at": "2023-01-01T00:00:00Z"
                                            }
                                        }
                                    },
                                    "limit": {
                                        "type": "integer",
                                        "format": "int32",
                                        "minimum": 0
                                    },
                                    "page": {
                                        "type": "integer",
                                        "format": "int32",
                                        "minimum": 0
                                    },
                                    "total": {
                                        "type": "integer",
                                        "format": "int64",
                                        "minimum": 0
                                    },
                                    "total_pages": {
                                        "type": "integer",
                                        "format": "int32",
                                        "minimum": 0
                                    }
                                }
                            }
                        ],
                        "nullable": true
                    },
                    "error": {
                        "type": "string",
                        "nullable": true
                    },
                    "message": {
                        "type": "string",
                        "nullable": true
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
            "ApiResponseRecommendationResponse": {
                "type": "object",
                "description": "Standard API response wrapper",
                "required": [
                    "success"
                ],
                "properties": {
                    "data": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/RecommendationResponse"
                            }
                        ],
                        "nullable": true
                    },
                    "error": {
                        "type": "string",
                        "nullable": true
                    },
                    "message": {
                        "type": "string",
                        "nullable": true
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
            "CreateRecommendationRequest": {
                "type": "object",
                "required": [
                    "dimension_key",
                    "priority",
                    "description"
                ],
                "properties": {
                    "description": {
                        "type": "string",
                        "description": "Detailed description of the recommendation",
                        "example": "Implement automated testing for critical components"
                    },
                    "dimension_id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Language-specific dimension ID (resolved by backend from dimension_key + language)",
                        "nullable": true
                    },
                    "dimension_key": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Stable cross-language dimension identifier",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "language": {
                        "type": "string",
                        "description": "Language code for this content entry (e.g. en, fr, pt, ss)",
                        "default": "en"
                    },
                    "priority": {
                        "$ref": "#/components/schemas/RecommendationPriority"
                    }
                },
                "example": {
                    "dimension_key": "550e8400-e29b-41d4-a716-446655440000",
                    "priority": "HIGH",
                    "description": "Implement automated testing for critical components",
                    "language": "en"
                }
            },
...
            "RecommendationPriority": {
                "type": "string",
                "enum": [
                    "LOW",
                    "MEDIUM",
                    "HIGH"
                ],
                "example": "HIGH"
            },
...
            "UpdateRecommendationRequest": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string",
                        "description": "New description (optional)",
                        "example": "Updated recommendation with more details",
                        "nullable": true
                    },
                    "language": {
                        "type": "string",
                        "description": "Language code for this content entry (e.g. en, fr, pt, ss)",
                        "nullable": true
                    },
                    "priority": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/RecommendationPriority"
                            }
                        ],
                        "nullable": true
                    }
                }
            },
...
        }
    },
...
}